Separate custom room form and table from dialog
