+++ client/src/components/RoomSelectionDialog.tsx
@@ -3,33 +3,16 @@ import logo from '../assets/logo.png'
 import styled from 'styled-components'
 import Button from '@mui/material/Button'
 import IconButton from '@mui/material/IconButton'
-import Paper from '@mui/material/Paper'
-import TableContainer from '@mui/material/TableContainer'
-import Table from '@mui/material/Table'
-import TableRow from '@mui/material/TableRow'
-import TableCell from '@mui/material/TableCell'
-import TableHead from '@mui/material/TableHead'
-import TableBody from '@mui/material/TableBody'
 import Tooltip from '@mui/material/Tooltip'
-import TextField from '@mui/material/TextField'
-import InputAdornment from '@mui/material/InputAdornment'
 import LinearProgress from '@mui/material/LinearProgress'
-import Dialog from '@mui/material/Dialog'
-import DialogActions from '@mui/material/DialogActions'
-import DialogContent from '@mui/material/DialogContent'
 import Alert from '@mui/material/Alert'
 import Snackbar from '@mui/material/Snackbar'
-import Avatar from '@mui/material/Avatar'
 import HelpOutlineIcon from '@mui/icons-material/HelpOutline'
 import ArrowBackIcon from '@mui/icons-material/ArrowBack'
-import PeopleAltIcon from '@mui/icons-material/PeopleAlt'
-import LockIcon from '@mui/icons-material/Lock'
-import Visibility from '@mui/icons-material/Visibility'
-import VisibilityOff from '@mui/icons-material/VisibilityOff'
 
-import { IRoomData } from '../../../types/Rooms'
+import { CustomRoomTable } from './CustomRoomTable'
+import { CreateRoomForm } from './CreateRoomForm'
 import { useAppSelector } from '../hooks'
-import { getAvatarString, getColorByString } from '../util'
 
 import phaserGame from '../PhaserGame'
 import Bootstrap from '../scenes/Bootstrap'
@@ -77,13 +60,6 @@ const Title = styled.h1`
   text-align: center;
 `
 
-const MessageText = styled.p`
-  margin: 10px;
-  font-size: 18px;
-  color: #eee;
-  text-align: center;
-`
-
 const Content = styled.div`
   display: flex;
   flex-direction: column;
@@ -95,54 +71,6 @@ const Content = styled.div`
   img {
     border-radius: 8px;
     height: 120px;
-  }
-`
-
-const CustomRoomTableContainer = styled(TableContainer)`
-  max-height: 500px;
-
-  table {
-    min-width: 650px;
-  }
-`
-
-const TableRowWrapper = styled(TableRow)`
-  &:last-child td,
-  &:last-child th {
-    border: 0;
-  }
-
-  .lock-icon {
-    font-size: 18px;
-  }
-
-  .name {
-    min-width: 100px;
-    overflow-wrap: anywhere;
-  }
-
-  .description {
-    min-width: 200px;
-    overflow-wrap: anywhere;
-  }
-`
-
-const CreateRoomFormWrapper = styled.form`
-  display: flex;
-  flex-direction: column;
-  width: 320px;
-  gap: 20px;
-`
-
-const PasswordDialog = styled(Dialog)`
-  .dialog-content {
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-  }
-
-  .MuiDialog-paper {
-    background: #222639;
   }
 `
 
@@ -159,230 +87,6 @@ const ProgressBarWrapper = styled.div`
 const ProgressBar = styled(LinearProgress)`
   width: 360px;
 `
-
-const CustomRoomTable = () => {
-  const [password, setPassword] = useState('')
-  const [selectedRoom, setSelectedRoom] = useState('')
-  const [showPasswordDialog, setShowPasswordDialog] = useState(false)
-  const [showPasswordError, setShowPasswordError] = useState(false)
-  const [passwordFieldEmpty, setPasswordFieldEmpty] = useState(false)
-  const lobbyJoined = useAppSelector((state) => state.room.lobbyJoined)
-  const availableRooms = useAppSelector((state) => state.room.availableRooms)
-
-  const handleJoinClick = (roomId: string, password: string | null) => {
-    if (!lobbyJoined) return
-    const bootstrap = phaserGame.scene.keys.bootstrap as Bootstrap
-    bootstrap.network
-      .joinCustomById(roomId, password)
-      .then(() => bootstrap.launchGame())
-      .catch((error) => {
-        console.log(error)
-        if (password) setShowPasswordError(true)
-      })
-  }
-
-  const handlePasswordSubmit = (event: React.FormEvent<HTMLFormElement>) => {
-    event.preventDefault()
-    const isValidPassword = password !== ''
-
-    if (isValidPassword === passwordFieldEmpty) setPasswordFieldEmpty(!passwordFieldEmpty)
-    if (isValidPassword) handleJoinClick(selectedRoom, password)
-  }
-
-  const resetPasswordDialog = () => {
-    setShowPasswordDialog(false)
-    setPassword('')
-    setPasswordFieldEmpty(false)
-    setShowPasswordError(false)
-  }
-
-  return availableRooms.length === 0 ? (
-    <MessageText>There are no custom rooms now, create one or join the public lobby.</MessageText>
-  ) : (
-    <>
-      <CustomRoomTableContainer component={Paper}>
-        <Table>
-          <TableHead>
-            <TableRow>
-              <TableCell></TableCell>
-              <TableCell>Name</TableCell>
-              <TableCell>Description</TableCell>
-              <TableCell>ID</TableCell>
-              <TableCell align="center">
-                <PeopleAltIcon />
-              </TableCell>
-              <TableCell align="right"></TableCell>
-            </TableRow>
-          </TableHead>
-          <TableBody>
-            {availableRooms.map((room) => {
-              const { roomId, metadata, clients } = room
-              const { name, description, hasPassword } = metadata
-              return (
-                <TableRowWrapper key={roomId}>
-                  <TableCell>
-                    <Avatar
-                      sx={{ height: '24px', width: '24px', fontSize: 15 }}
-                      style={{ background: getColorByString(name) }}
-                    >
-                      {getAvatarString(name)}
-                    </Avatar>
-                  </TableCell>
-                  <TableCell>
-                    <div className="name">{name}</div>
-                  </TableCell>
-                  <TableCell>
-                    <div className="description">{description}</div>
-                  </TableCell>
-                  <TableCell>{roomId}</TableCell>
-                  <TableCell align="center">{clients}</TableCell>
-                  <TableCell align="center">
-                    <Tooltip title={hasPassword ? 'Password required' : ''}>
-                      <Button
-                        variant="outlined"
-                        color="secondary"
-                        onClick={() => {
-                          if (hasPassword) {
-                            setShowPasswordDialog(true)
-                            setSelectedRoom(roomId)
-                          } else {
-                            handleJoinClick(roomId, null)
-                          }
-                        }}
-                      >
-                        <div style={{ display: 'flex', gap: 3 }}>
-                          {hasPassword && <LockIcon className="lock-icon" />}
-                          Join
-                        </div>
-                      </Button>
-                    </Tooltip>
-                  </TableCell>
-                </TableRowWrapper>
-              )
-            })}
-          </TableBody>
-        </Table>
-      </CustomRoomTableContainer>
-      <PasswordDialog open={showPasswordDialog} onClose={resetPasswordDialog}>
-        <form onSubmit={handlePasswordSubmit}>
-          <DialogContent className="dialog-content">
-            <MessageText>This a private room, please enter password:</MessageText>
-            <TextField
-              autoFocus
-              fullWidth
-              error={passwordFieldEmpty}
-              helperText={passwordFieldEmpty && 'Required'}
-              value={password}
-              label="Password"
-              type="password"
-              variant="outlined"
-              color="secondary"
-              onInput={(e) => {
-                setPassword((e.target as HTMLInputElement).value)
-              }}
-            />
-            {showPasswordError && (
-              <Alert severity="error" variant="outlined">
-                Incorrect Password!
-              </Alert>
-            )}
-          </DialogContent>
-          <DialogActions>
-            <Button color="secondary" onClick={resetPasswordDialog}>
-              Cancel
-            </Button>
-            <Button color="secondary" type="submit">
-              Join
-            </Button>
-          </DialogActions>
-        </form>
-      </PasswordDialog>
-    </>
-  )
-}
-
-const CreateRoomForm = () => {
-  const [values, setValues] = useState<IRoomData>({
-    name: '',
-    description: '',
-    password: null,
-    autoDispose: true,
-  })
-  const [showPassword, setShowPassword] = useState(false)
-  const [nameFieldEmpty, setNameFieldEmpty] = useState(false)
-  const [descriptionFieldEmpty, setDescriptionFieldEmpty] = useState(false)
-  const lobbyJoined = useAppSelector((state) => state.room.lobbyJoined)
-
-  const handleChange = (prop: keyof IRoomData) => (event: React.ChangeEvent<HTMLInputElement>) => {
-    setValues({ ...values, [prop]: event.target.value })
-  }
-  const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
-    event.preventDefault()
-    const isValidName = values.name !== ''
-    const isValidDescription = values.description !== ''
-
-    if (isValidName === nameFieldEmpty) setNameFieldEmpty(!nameFieldEmpty)
-    if (isValidDescription === descriptionFieldEmpty)
-      setDescriptionFieldEmpty(!descriptionFieldEmpty)
-
-    // create custom room if name and description are not empty, else show error
-    if (isValidName && isValidDescription && lobbyJoined) {
-      const bootstrap = phaserGame.scene.keys.bootstrap as Bootstrap
-      bootstrap.network
-        .createCustom(values)
-        .then(() => bootstrap.launchGame())
-        .catch((error) => console.log(error))
-    }
-  }
-
-  return (
-    <CreateRoomFormWrapper onSubmit={handleSubmit}>
-      <TextField
-        label="Name"
-        variant="outlined"
-        color="secondary"
-        autoFocus
-        error={nameFieldEmpty}
-        helperText={nameFieldEmpty && 'Name is required'}
-        onChange={handleChange('name')}
-      />
-
-      <TextField
-        label="Description"
-        variant="outlined"
-        color="secondary"
-        error={descriptionFieldEmpty}
-        helperText={descriptionFieldEmpty && 'Description is required'}
-        multiline
-        rows={4}
-        onChange={handleChange('description')}
-      />
-
-      <TextField
-        type={showPassword ? 'text' : 'password'}
-        label="Password (optional)"
-        onChange={handleChange('password')}
-        color="secondary"
-        InputProps={{
-          endAdornment: (
-            <InputAdornment position="end">
-              <IconButton
-                aria-label="toggle password visibility"
-                onClick={() => setShowPassword(!showPassword)}
-                edge="end"
-              >
-                {showPassword ? <VisibilityOff /> : <Visibility />}
-              </IconButton>
-            </InputAdornment>
-          ),
-        }}
-      />
-      <Button variant="contained" color="secondary" type="submit">
-        Create
-      </Button>
-    </CreateRoomFormWrapper>
-  )
-}
 
 export default function RoomSelectionDialog() {
   const [showCustomRoom, setShowCustomRoom] = useState(false)
